(pup-2316) Force aix user attributes to be an array
Prior to this change, if the attributes parameters was specified
as a single element array, it would get flattened to a string
(by copy_as_resource I believe) resulting in an exception when
the attributes are accessed with the map method.

This change follows a long tradition of array-ifying elements
that are not an array.